refactored cacheinstance init code to make it clean, added old fibercachemanagersuite logic back to test when disable cache separation
